+++ validate-cli/pom.xml
@@ -1,60 +1,23 @@
 <?xml version="1.0"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <artifactId>validering</artifactId>
-    <groupId>no.difi.vefa</groupId>
-    <version>1.0.0</version>
-    <relativePath>..</relativePath>
-  </parent>
-  <artifactId>validate-cli</artifactId>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+         xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <artifactId>validering</artifactId>
+        <groupId>no.difi.vefa</groupId>
+        <version>1.0.0</version>
+        <relativePath>..</relativePath>
+    </parent>
 
-  <build>
-  	<finalName>validate-cli</finalName>
-  	<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>2.3.2</version>
-				<configuration>
-					<source>1.6</source>
-					<target>1.6</target>
-				</configuration>
-			</plugin>
-			
-			<plugin>
-			    <artifactId>maven-assembly-plugin</artifactId>
-			    <version>2.3</version>
-			    <configuration>
-			        <appendAssemblyId>false</appendAssemblyId>
-			        <archive>
-			            <manifest>
-			                <mainClass>no.difi.vefa.cli.App</mainClass>
-			            </manifest>
-			        </archive>
-			        <descriptorRefs>
-			            <descriptorRef>jar-with-dependencies</descriptorRef>
-			        </descriptorRefs>
-			    </configuration>
-			    <executions>
-			        <execution>
-			            <id>make-assembly</id>
-			            <phase>package</phase>
-			            <goals>
-			                <goal>single</goal>
-			            </goals>
-			        </execution>
-			    </executions>
-			</plugin>			  	
-  	</plugins>
-  </build>
-    
-  <dependencies>
-    <dependency>
-    	<groupId>no.difi.vefa</groupId>
-    	<artifactId>validate-lib</artifactId>
-    	<version>1.0.0</version>
-    </dependency>
-  </dependencies>
+    <artifactId>validate-cli</artifactId>
+    <packaging>jar</packaging>
+
+    <dependencies>
+        <dependency>
+            <groupId>no.difi.vefa</groupId>
+            <artifactId>validate-lib</artifactId>
+            <version>1.0.0</version>
+        </dependency>
+    </dependencies>
 </project>
