+++ validate-lib/src/main/java/no/difi/vefa/cache/XSLTransformerCache.java
@@ -22,41 +22,37 @@ public class XSLTransformerCache {
      */
     private Ehcache transformerCache;
 
-    public XSLTransformerCache()
-    {
+    public XSLTransformerCache() {
         // Load our Transformer cache:
-    	transformerCache = cacheManager.getEhcache( "transformer" );
+        transformerCache = cacheManager.getEhcache("transformer");
     }
 
     /**
      * Adds a new Templates object to the Cache
      *
      * @param xslFileName The name of the xsl file
-     * @param templates The thread safe templates object which contains the parsed version of the xslFile
+     * @param templates   The thread safe templates object which contains the parsed version of the xslFile
      */
-    public void addTemplate(String xslFileName, Templates templates)
-    {
+    public void addTemplate(String xslFileName, Templates templates) {
         // Create an EHCache Element to hold the widget
-        Element templatesElement = new Element( xslFileName, templates );
+        Element templatesElement = new Element(xslFileName, templates);
 
         // Add the element to the cache
-        transformerCache.put( templatesElement );
+        transformerCache.put(templatesElement);
     }
 
     /**
      * Retrieves a Templates object from the cache
      *
-     * @param xslFileName   The name of the xsl file to retrieve
+     * @param xslFileName The name of the xsl file to retrieve
      * @return The Templates object containing the parsed xslFile, or null if not in the cache.
      */
     public Templates getTemplate(String xslFileName) {
         // Retrieve the element that contains the requested Templates object
-        Element element = transformerCache.get( xslFileName );
-        if( element != null )
-        {
-        	// Using getObjectValue instead of getValue since Templates object is not serializable.
-            Templates objectValue = (Templates) element.getObjectValue();
-            return objectValue;
+        Element element = transformerCache.get(xslFileName);
+        if (element != null) {
+            // Using getObjectValue instead of getValue since Templates object is not serializable.
+            return (Templates) element.getObjectValue();
         }
 
         // We don't have the object in the cache so return null
