+++ validate-lib/src/main/java/no/difi/vefa/validation/XSDValidation.java
@@ -1,12 +1,9 @@
 package no.difi.vefa.validation;
 
-import no.difi.vefa.message.Message;
 import no.difi.vefa.message.Messages;
-import no.difi.vefa.properties.PropertiesFile;
+import no.difi.vefa.util.PropertiesUtils;
 import no.difi.vefa.util.MessageUtils;
 import org.w3c.dom.Document;
-
-import java.util.List;
 
 /**
  * This class can be used to check if an XML Document
@@ -21,12 +18,12 @@ public class XSDValidation {
      * @param xmlDoc         XML as Document
      * @param xsdFile        Path to XSD file as String
      * @param messages       List of messages
-     * @param propertiesFile PropertiesFile
+     * @param propertiesUtils PropertiesFile
      */
-    public void main(Document xmlDoc, String xsdFile, Messages messages, PropertiesFile propertiesFile) {
+    public void main(Document xmlDoc, String xsdFile, Messages messages, PropertiesUtils propertiesUtils) {
         try {
-            no.difi.vefa.xml.XSDValidation xsdValidation = new no.difi.vefa.xml.XSDValidation();
-            xsdValidation.main(xmlDoc, xsdFile, propertiesFile);
+            no.difi.vefa.util.xml.XSDValidation xsdValidation = new no.difi.vefa.util.xml.XSDValidation();
+            xsdValidation.main(xmlDoc, xsdFile, propertiesUtils);
         } catch (Exception e) {
             messages.addMessage(MessageUtils.translate(e));
         }
