+++ validate-lib/src/main/resources/ehcache.xml
@@ -1,43 +1,43 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <ehcache name="Default">
-  <defaultCache
-      maxElementsInMemory="10000"
-      eternal="false"
-      timeToIdleSeconds="1200"
-      timeToLiveSeconds="1200"
-      overflowToDisk="true"
-      diskSpoolBufferSizeMB="30"
-      maxElementsOnDisk="10000000"
-      diskPersistent="false"
-      diskExpiryThreadIntervalSeconds="120"
-      memoryStoreEvictionPolicy="LRU"/>
+    <defaultCache
+            maxElementsInMemory="10000"
+            eternal="false"
+            timeToIdleSeconds="1200"
+            timeToLiveSeconds="1200"
+            overflowToDisk="true"
+            diskSpoolBufferSizeMB="30"
+            maxElementsOnDisk="10000000"
+            diskPersistent="false"
+            diskExpiryThreadIntervalSeconds="120"
+            memoryStoreEvictionPolicy="LRU"/>
 
-  <cache name="transformer"
-         maxElementsInMemory="100"
-         maxElementsOnDisk="0"
-         eternal="false"
-         timeToIdleSeconds="120"
-         timeToLiveSeconds="0"
-         memoryStoreEvictionPolicy="LFU">
-  </cache>
+    <cache name="transformer"
+           maxElementsInMemory="100"
+           maxElementsOnDisk="0"
+           eternal="false"
+           timeToIdleSeconds="120"
+           timeToLiveSeconds="0"
+           memoryStoreEvictionPolicy="LFU">
+    </cache>
 
-  <cache name="properties"
-         maxElementsInMemory="1"
-         maxElementsOnDisk="0"
-         eternal="false"
-         timeToIdleSeconds="120"
-         timeToLiveSeconds="0"
-         memoryStoreEvictionPolicy="LFU">
-  </cache>
+    <cache name="properties"
+           maxElementsInMemory="1"
+           maxElementsOnDisk="0"
+           eternal="false"
+           timeToIdleSeconds="120"
+           timeToLiveSeconds="0"
+           memoryStoreEvictionPolicy="LFU">
+    </cache>
 
-  <cache name="configuration"
-         maxElementsInMemory="5"
-         maxElementsOnDisk="0"
-         eternal="false"
-         timeToIdleSeconds="120"
-         timeToLiveSeconds="0"
-         memoryStoreEvictionPolicy="LFU">
-  </cache>
-    
+    <cache name="configuration"
+           maxElementsInMemory="5"
+           maxElementsOnDisk="0"
+           eternal="false"
+           timeToIdleSeconds="120"
+           timeToLiveSeconds="0"
+           memoryStoreEvictionPolicy="LFU">
+    </cache>
+
 </ehcache>
