+++ validate-lib/src/test/java/no/difi/vefa/configuration/ConfigurationTest.java
@@ -2,7 +2,7 @@ package no.difi.vefa.configuration;
 
 import static org.junit.Assert.*;
 
-import no.difi.vefa.properties.PropertiesFile;
+import no.difi.vefa.util.PropertiesUtils;
 
 import org.junit.Test;
 import org.w3c.dom.Document;
@@ -12,7 +12,7 @@ public class ConfigurationTest {
 	@Test
 	public void testFileToXMLDOM() throws Exception {
 		String path = new java.io.File("src/test/resources/validator.properties").getCanonicalPath();
-		PropertiesFile propFile = new PropertiesFile();
+		PropertiesUtils propFile = new PropertiesUtils();
 		propFile.main(path);
 		propFile.dataDir = new java.io.File(".").getCanonicalPath();
