+++ validate-lib/src/test/java/no/difi/vefa/validation/XSDValidationTest.java
@@ -1,18 +1,15 @@
 package no.difi.vefa.validation;
 
-import no.difi.vefa.message.Message;
 import no.difi.vefa.message.MessageType;
 import no.difi.vefa.message.Messages;
 import no.difi.vefa.message.ValidationType;
-import no.difi.vefa.properties.PropertiesFile;
-import no.difi.vefa.xml.Utils;
+import no.difi.vefa.util.PropertiesUtils;
+import no.difi.vefa.util.xml.XmlUtils;
 import org.junit.Before;
 import org.junit.Test;
 import org.w3c.dom.Document;
 
 import java.io.File;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Scanner;
 
 import static org.junit.Assert.assertEquals;
@@ -20,21 +17,21 @@ import static org.junit.Assert.assertEquals;
 public class XSDValidationTest {
 
     private String basePath;
-    private Utils utils;
+    private XmlUtils xmlUtils;
     private String xml;
     private Document xmlDoc;
     private Messages messages;
     private XSDValidation xsdValidation;
-    private PropertiesFile propFile;
+    private PropertiesUtils propFile;
     private Validate validate;
 
     @Before
     public void setUp() throws Exception {
         basePath = new java.io.File("src/test/resources/").getCanonicalPath();
-        utils = new Utils();
+        xmlUtils = new XmlUtils();
         xsdValidation = new XSDValidation();
         validate = new Validate();
-        propFile = validate.getPropertiesFile();
+        propFile = validate.getPropertiesUtils();
     }
 
     @Test
@@ -43,14 +40,14 @@ public class XSDValidationTest {
 
         scanner = new Scanner(new File(basePath + "/TestXSD.xml"));
         xml = scanner.useDelimiter("\\Z").next();
-        xmlDoc = utils.stringToXMLDOM(xml);
+        xmlDoc = xmlUtils.stringToXMLDOM(xml);
         messages = new Messages();
         xsdValidation.main(xmlDoc, basePath + "/TestXSD.xsd", messages, propFile);
         assertEquals(0, messages.getMessages().size());
 
         scanner = new Scanner(new File(basePath + "/TestXSDWithErrors.xml"));
         xml = scanner.useDelimiter("\\Z").next();
-        xmlDoc = utils.stringToXMLDOM(xml);
+        xmlDoc = xmlUtils.stringToXMLDOM(xml);
         messages = new Messages();
         xsdValidation.main(xmlDoc, basePath + "/TestXSD.xsd", messages, propFile);
         assertEquals(1, messages.getMessages().size());
