+++ validate-ws/pom.xml
@@ -9,7 +9,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>validate-soap</artifactId>
+    <artifactId>validate-ws</artifactId>
     <packaging>jar</packaging>
 
     <dependencies>
