+++ validate-ws/src/test/java/no/difi/vefa/ws/ListVersionsTest.java
@@ -2,8 +2,8 @@ package no.difi.vefa.ws;
 
 import static org.junit.Assert.*;
 
-import no.difi.vefa.properties.PropertiesFile;
-import no.difi.vefa.ws.rest.ListVersions;
+import no.difi.vefa.util.PropertiesUtils;
+import no.difi.vefa.ws.rest.model.ListVersions;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -14,13 +14,13 @@ public class ListVersionsTest {
 	@Before
 	public void setUp() throws Exception {
 		String path = new java.io.File("src/test/resources/validator.properties").getCanonicalPath();
-		PropertiesFile propFile = new PropertiesFile();
+		PropertiesUtils propFile = new PropertiesUtils();
 		propFile.main(path);
 		propFile.dataDir = new java.io.File("src/test/resources/").getCanonicalPath();
 		
 		listVersions = new ListVersions();
 		listVersions.baseUri = "http://www.test.com/validate-ws/";
-		listVersions.propertiesFile = propFile;
+		listVersions.propertiesUtils = propFile;
 	}
 
 	@Test
